+++ app/src/main/java/com/lewgmail/romanenko/taxiservice/view/ViewDriver/OrderInf.java
@@ -95,7 +95,7 @@ public class OrderInf extends AppCompatActivity implements EditOrderInterface {
         progress.show();
         // sdf = intentMy.getStringExtra("keyNumberOfOrder");
         // lol = Integer.parseInt(intentMy.getStringExtra("keyNumberOfOrder"));
-        loadOrderForDriverPresenter.loadOrderSpecificId(Integer.parseInt(intentMy.getStringExtra("keyNumberOfOrder")));
+        loadOrderForDriverPresenter.loadOrderSpecificId(Long.parseLong(intentMy.getStringExtra("keyNumberOfOrder")));
 
 
     }
